+++ src/common/components/BoxText/SnsLink/SnsLink.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import FeatureTabIcons from '@icons/FeatureTabIcons/FeatureTabIcons';
 import { FeatureTabIconsState } from '@icons/FeatureTabIcons/FeatureTabIcons.types';
