+++ src/common/components/BoxText/SurveyInfo/SurveyInfo.server.tsx
